I'm in the process of implementing the sort_common() function. The algorithm for finding the shortest sequences of operations to move elements from stack A into their correct positions in stack B is almost complete, but testing still needs to be done. Some 42 headers have been added

+++ stack_basis.h
@@ -19,15 +19,15 @@ typedef struct	s_stack
 	size_t	capacity;
 }	t_stack;
 
-int		stack_init(t_stack *stack, const size_t capacity);
-void	stack_free(t_stack *stack);
-void	stack_print(t_stack *stack);
-void	stack_push(t_stack *stack, const int elem);
-int		*stack_pop(t_stack *stack);
-int		stack_empty(t_stack *stack);
-int		stack_sorted(t_stack *stack);
-int		stack_contains(t_stack *stack, int elem);
-int		stack_copy(t_stack *dst, t_stack *src);
-size_t	stack_get_elem_index(t_stack *stack, int elem);
+int			stack_init(t_stack *stack, const size_t capacity);
+void		stack_free(t_stack *stack);
+void		stack_print(t_stack *stack);
+void		stack_push(t_stack *stack, const int elem);
+int			*stack_pop(t_stack *stack);
+int			stack_empty(t_stack *stack);
+int			stack_sorted(t_stack *stack);
+int			stack_contains(t_stack *stack, int elem);
+int			stack_copy(t_stack *dst, t_stack *src);
+long long	stack_get_elem_index(t_stack *stack, int elem);
 
 #endif
